delete videoframes

+++ ib_aitool/admin/interview_analyzer/views.py
@@ -21,10 +21,8 @@ import os
 import jinja2
 import glob
 
-from ib_aitool.admin.interview_analyzer.generate_video_transcript import generate_transcipt, save_frames_for_timestamps, \
-    save_audioclip_timestamps, analyze_timestamp_folder, analyze_audio_timestamps_clips
-from ib_aitool.admin.interview_analyzer.save_video_analysis_data import save_videots_report, \
-    generate_and_save_overall_video_report
+from ib_aitool.admin.interview_analyzer.generate_video_transcript import generate_transcipt,save_frames_for_timestamps,save_audioclip_timestamps,analyze_timestamp_folder,analyze_audio_timestamps_clips
+from ib_aitool.admin.interview_analyzer.save_video_analysis_data import save_videots_report,generate_and_save_overall_video_report
 from jinja2 import Environment
 import shutil
 
@@ -208,43 +206,40 @@ def calculate_overall_confidence(facial_emotion_data):
     facial_emotion_data = facial_emotion_data.replace("'", "\"")
     facial_emotion_data = json.loads(facial_emotion_data)
 
-    # Define weights for selected emotions (neutral, happy, and surprise)
-    weight_neutral = 0.7
-    weight_happy = 0.2
-    weight_surprise = 0.05
-    weight_angry = 0.6
-    weight_fear = 0.7
-    weight_sad = 1
-    weight_disgust = 1
+    neutral_percentage = facial_emotion_data['neutral'] * 100
+    happy_percentage = facial_emotion_data['happy'] * 100
+    fear_percentage = facial_emotion_data['fear'] * 100
+    angry_percentage = facial_emotion_data['angry'] * 100
+    sad_percentage = facial_emotion_data['sad'] * 100
+    surprise_percentage = facial_emotion_data['surprise'] * 100
 
-    # Calculate the weighted sum of selected emotions
-    weighted_sum = (
-            facial_emotion_data['neutral'] * weight_neutral +
-            facial_emotion_data['happy'] * weight_happy +
-            facial_emotion_data['surprise'] * weight_surprise
-    )
+    if neutral_percentage >= 70 and happy_percentage >= 20:
+        weighted_average = 100 - (fear_percentage + angry_percentage + sad_percentage)
+    elif (
+            happy_percentage <= 20
+            and neutral_percentage <= 70
+            and angry_percentage > 5
+            and fear_percentage > 5
+    ):
+        weighted_average = 100 - (fear_percentage + angry_percentage + sad_percentage)
+    elif angry_percentage >= 40 and fear_percentage >= 30:
+        weighted_average = 100 - (fear_percentage + angry_percentage + sad_percentage)
+    elif neutral_percentage >= 40 and happy_percentage >= 5:
+        weighted_average = 100 - (fear_percentage + angry_percentage + sad_percentage)
+    else:
+        weighted_average = 55
+    weighted_average = weighted_average + surprise_percentage
 
-    # Calculate the sum of remaining emotions (angry, disgust, fear, sad)
-    other_emotions_sum = (
-            facial_emotion_data['angry'] * weight_angry +
-            facial_emotion_data['disgust'] * weight_disgust +
-            facial_emotion_data['fear'] * weight_fear +
-            facial_emotion_data['sad'] * weight_sad
-    )
+    CS = (happy_percentage + neutral_percentage+surprise_percentage)
 
-    # Calculate the nervousness score (sum of remaining emotions)
-    nervousness_score = other_emotions_sum
+    NS = fear_percentage + sad_percentage
+    CL = CS / (CS + NS)
 
-    # Calculate the confidence score (weighted sum of selected emotions)
-    confidence_score = weighted_sum
-
-    # Normalize the scores to percentages
-    total_score = confidence_score + nervousness_score
-    CL = (confidence_score / total_score) * 100
-    NS = (nervousness_score / total_score) * 100
-    # CS = ((facial_emotion_data['happy']*100) + (facial_emotion_data['neutral']*100)+(facial_emotion_data['surprise']*100))
-
-    return 0, 0, NS, CL
+    # weighted_average= weighted_average+ facial_emotion_data['surprise']
+    # Subtract the percentages of 'angry' and 'fear' emotions
+    # Ensure that the overall confidence is within the range of 0% to 100%
+    overall_confidence = max(0, min(weighted_average, 100))
+    return overall_confidence, CS, NS, CL
 
 
 def generate_report_pdf(candidate_id):
@@ -286,40 +281,22 @@ def create_overall_data_by_candidate_id(candidate_id):
 
     # Create dictionaries to store the values
     interviewer_confidence_dict = {}
-    interviewer_confidence_dict_text = {}
     candidate_confidence_dict = {}
-    candidate_confidence_dict_text = {}
 
-    # Calculate and store the values for the interviewer video
+    # Calculate and store the values for the interviewer
     overall_interviewer_confidence, CS, NS, CL = calculate_overall_confidence(
         candidate.overall_interviewer_video_report)
     interviewer_confidence_dict['overall_confidence'] = overall_interviewer_confidence
     interviewer_confidence_dict['CS'] = CS
     interviewer_confidence_dict['NS'] = NS
-    interviewer_confidence_dict['CL'] = CL
-
-    # Calculate and store the values for the interviewer text analysis
-    overall_interviewer_confidence_text, CS, NS, CL = calculate_overall_confidence(
-        candidate.overall_interviewer_text_report)
-    interviewer_confidence_dict_text['overall_confidence'] = overall_interviewer_confidence_text
-    interviewer_confidence_dict_text['CS'] = CS
-    interviewer_confidence_dict_text['NS'] = NS
-    interviewer_confidence_dict_text['CL'] = CL
+    interviewer_confidence_dict['CL'] = CL * 100
 
     # Calculate and store the values for the candidate
     overall_candidate_confidence, CS, NS, CL = calculate_overall_confidence(candidate.overall_candidate_video_report)
     candidate_confidence_dict['overall_confidence'] = overall_candidate_confidence
     candidate_confidence_dict['CS'] = CS
     candidate_confidence_dict['NS'] = NS
-    candidate_confidence_dict['CL'] = CL
-
-    # Calculate and store the values for the candidate text
-    overall_candidate_confidence_text, CS, NS, CL = calculate_overall_confidence(
-        candidate.overall_candidate_text_report)
-    candidate_confidence_dict_text['overall_confidence'] = overall_candidate_confidence_text
-    candidate_confidence_dict_text['CS'] = CS
-    candidate_confidence_dict_text['NS'] = NS
-    candidate_confidence_dict_text['CL'] = CL
+    candidate_confidence_dict['CL'] = CL * 100
 
     overall = {"candidate_id": str(candidate_id),
                "interviewer_video_report": ib_format_json(data=candidate.overall_interviewer_video_report),
@@ -330,8 +307,6 @@ def create_overall_data_by_candidate_id(candidate_id):
                "candidate_audio_report": ib_format_json(data=candidate.overall_candidate_audio_report),
                "overall_interviewer_confidence": interviewer_confidence_dict,
                "overall_candidate_confidence": candidate_confidence_dict,
-               "overall_interviewer_confidence_text": interviewer_confidence_dict_text,
-               "overall_candidate_confidence_text": candidate_confidence_dict_text,
                }
 
     data = get_video_data(candidate_id)
@@ -624,6 +599,7 @@ def run_tasks():
                         db.session.commit()
                     c_data = Candidate.get_video_data(candidate_id)
                     remove_all_model_created_files(c_data.interview_video)
+
             else:
                 final_result = False
     else:
@@ -649,7 +625,12 @@ def remove_all_model_created_files(videopath):
 
     # List all files in the folder
     all_files = glob.glob(os.path.join(BASE_DIR, '*'))
-
+    
+    video_name = os.path.basename(video_file_path)
+    # Remove the file extension if needed
+    video_name_without_extension, extension = os.path.splitext(video_name)
+    video_name_without_extension = f'uploads/{video_name_without_extension}'
+    shutil.rmtree(video_name_without_extension)
     # Iterate through the files and delete those with matching base names
     for file_path in all_files:
         print(file_path)
@@ -672,52 +653,36 @@ def get_video_data(video_id):
         print(f"Error: {e}")
         return None
 
-
 def calculate_qna_confidence(facial_emotion_data):
+
     facial_emotion_data = facial_emotion_data.replace("'", "\"")
     facial_emotion_data = json.loads(facial_emotion_data)
 
-    weight_neutral = 0.7
-    weight_happy = 0.2
-    weight_surprise = 0.05
-    weight_angry = 0.6
-    weight_fear = 0.7
-    weight_sad = 1
-    weight_disgust = 1
+    neutral_percentage = facial_emotion_data['neutral'] * 100
+    happy_percentage = facial_emotion_data['happy'] * 100
+    fear_percentage = facial_emotion_data['fear'] * 100
+    angry_percentage = facial_emotion_data['angry'] * 100
+    sad_percentage = facial_emotion_data['sad'] * 100
+    surprise_percentage = facial_emotion_data['surprise'] * 100
+    CS = (happy_percentage + neutral_percentage+surprise_percentage)
+    NS = fear_percentage + sad_percentage
+    CL = CS / (CS + NS)  
+    
+    if CS <0:
+        CS=0
+    if NS <0:
+        NS=0
+    if  CL<0:
+        CL=0 
 
-    # Calculate the weighted sum of selected emotions
-    weighted_sum = (
-            facial_emotion_data['neutral'] * weight_neutral +
-            facial_emotion_data['happy'] * weight_happy +
-            facial_emotion_data['surprise'] * weight_surprise
-    )
-
-    # Calculate the sum of remaining emotions (angry, disgust, fear, sad)
-    other_emotions_sum = (
-            facial_emotion_data['angry'] * weight_angry +
-            facial_emotion_data['disgust'] * weight_disgust +
-            facial_emotion_data['fear'] * weight_fear +
-            facial_emotion_data['sad'] * weight_sad
-    )
-
-    # Calculate the nervousness score (sum of remaining emotions)
-    nervousness_score = other_emotions_sum
-
-    # Calculate the confidence score (weighted sum of selected emotions)
-    confidence_score = weighted_sum
-
-    # Normalize the scores to percentages
-    total_score = confidence_score + nervousness_score
-    CL = (confidence_score / total_score) * 100
-    NS = (nervousness_score / total_score) * 100
-    # CS = ((facial_emotion_data['happy']*100) + (facial_emotion_data['neutral']*100)+(facial_emotion_data['surprise']*100))
-
-    return f"CS:{round(CL, 2)}%, NS:{round(NS, 2)}%"
-
+    CL=round(CL* 100, 2) 
+    # weighted_average= weighted_average+ facial_emotion_data['surprise']
+    # Subtract the percentages of 'angry' and 'fear' emotions
+    # Ensure that the overall confidence is within the range of 0% to 100%
+    return f"CS:{round(CS,2)}%, NS:{round(NS,2)}%, CL:{CL}%"
 
 # Register the custom Jinja2 filter
 app.jinja_env.filters['emotion_scores'] = calculate_qna_confidence
-
 
 @app.route('/delete_route/<int:item_id>')
 def delete_route(item_id):
@@ -741,7 +706,8 @@ def delete_route(item_id):
             os.remove(video_url)
             os.remove(video_pdf)
             os.remove(video_audio)
-            shutil.rmtree(video_name_without_extension)
+            if os.path.exists(video_name_without_extension):
+                shutil.rmtree(video_name_without_extension)
 
         db.session.commit()
 
@@ -751,7 +717,6 @@ def delete_route(item_id):
         return f"Error: {str(e)}", 500
 
     return redirect(url_for('interview_analyzer.index'))
-
 
 app.register_blueprint(
     products_blueprint, url_prefix='/smart-interview-assessment')
